+++ tests/acceptance/amendments/SupporterMinWarningCept.php
@@ -6,10 +6,22 @@ $I->populateDBData1();
 
 $I->wantTo('make sure the supporter-warning appears for natural persons');
 
-$page = $I->gotoConsultationHome(true, 'bdk', 'bdk');
-$I->loginAsStdAdmin();
-$I->click('.motionLink4');
+$page = $I->loginAndGotoStdAdminPage()->gotoMotionTypes(1);
+
+$I->dontSeeElement('section.amendmentSupporters');
+$I->seeCheckboxIsChecked('#sameInitiatorSettingsForAmendments input');
+$I->executeJS('$("#sameInitiatorSettingsForAmendments input").prop("checked", false).trigger("change");');
+$I->seeElement('section.amendmentSupporters');
+
+$I->selectFueluxOption('#typeSupportTypeAmendment', \app\models\supportTypes\SupportBase::GIVEN_BY_INITIATOR);
+$I->fillField('#typeMinSupportersAmendment', 19);
+
+$page->saveForm();
+
+$I->gotoMotion();
 $I->click('.sidebarActions .amendmentCreate a');
+
+$I->seeElement('.supporterData');
 
 $I->fillField('#initiatorPrimaryName', 'Mein Name');
 $I->fillField('#initiatorEmail', 'test@example.org');
@@ -36,3 +48,11 @@ $I->submitForm('#amendmentEditForm', [], 'save');
 $I->dontSeeBootboxDialog('Es müssen mindestens 19 Unterstützer*innen angegeben werden');
 $I->dontSee('Not enough supporters.');
 $I->see(mb_strtoupper('Änderungsantrag bestätigen'), 'h1');
+
+
+$I->wantTo('make sure the changes are not active for motions');
+
+$I->gotoConsultationHome();
+$I->click('#sidebar .createMotion1');
+
+$I->dontSeeElement('.supporterData');